Download manager update
git-svn-id: https://edges-3g-bridge.svn.sourceforge.net/svnroot/edges-3g-bridge/cancergrid/trunk/queue_manager@1875 a7169a2c-3604-0410-bc95-c702d8d87f7a

+++ DownloadManager.cpp
@@ -19,9 +19,9 @@
 /* For OpenSSL multithread support */
 static GMutex **ssl_mutexes;
 
-/* libcurl-related variables */
-static CURLSH *shared_data;
-static GStaticMutex shared_lock = G_STATIC_MUTEX_INIT;
+/* Garbage collector for DLItem */
+static GList *gc_list;
+static GStaticMutex gc_lock = G_STATIC_MUTEX_INIT;
 
 /**********************************************************************
  * libcurl thread-related callbacks for glib
@@ -29,12 +29,12 @@ static GStaticMutex shared_lock = G_STATIC_MUTEX_INIT;
 
 static void lock_curl(CURL *handle, curl_lock_data data, curl_lock_access access, void *userptr)
 {
-	g_static_mutex_lock(&shared_lock);
+	g_mutex_lock((GMutex *)userptr);
 }
 
 static void unlock_curl(CURL *handle, curl_lock_data data, void *userptr)
 {
-	g_static_mutex_unlock(&shared_lock);
+	g_mutex_unlock((GMutex *)userptr);
 }
 
 /**********************************************************************
@@ -83,75 +83,47 @@ static void ssl_dynlock_destroy(struct CRYPTO_dynlock_value *lock, const char *f
 }
 
 /**********************************************************************
- * The download handler function
+ * Garbage collector for class DLItem
  */
 
-static size_t write_callback(void *buf, size_t size, size_t items, void *data)
+static void add_to_gc(DLItem *item)
 {
-	DLItem *item = (DLItem *)data;
-	item->write(buf, size * items);
-	return size * items;
+	g_static_mutex_lock(&gc_lock);
+	gc_list = g_list_prepend(gc_list, item);
+	g_static_mutex_unlock(&gc_lock);
 }
 
-static void download_handler(void *data, void *user_data G_GNUC_UNUSED)
+static void run_gc(void)
 {
-	static GStaticPrivate key_curl = G_STATIC_PRIVATE_INIT;
-	static GStaticPrivate key_errbuf = G_STATIC_PRIVATE_INIT;
-	CURLcode result;
-	char *errbuf;
-	CURL *curl;
-
-	curl = g_static_private_get(&key_curl);
-	if (!curl)
+	g_static_mutex_lock(&gc_lock);
+	while (gc_list)
 	{
-		curl = curl_easy_init();
-		curl_easy_setopt(curl, CURLOPT_SHARE, shared_data);
-		errbuf = (char *)g_malloc(CURL_ERROR_SIZE);
-		curl_easy_setopt(curl, CURLOPT_ERRORBUFFER, errbuf);
-		curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_callback);
-		curl_easy_setopt(curl, CURLOPT_NOSIGNAL, (long)TRUE);
-		curl_easy_setopt(curl, CURLOPT_MAXREDIRS, 10);
-
-		/* XXX For debugging */
-		// curl_easy_setopt(curl, CURLOPT_VERBOSE, (long)TRUE);
-
-		g_static_private_set(&key_curl, curl, curl_easy_cleanup);
-		g_static_private_set(&key_errbuf, errbuf, g_free);
+		DLItem *item = (DLItem *)gc_list->data;
+		gc_list = g_list_delete_link(gc_list, gc_list);
+		delete item;
 	}
-	errbuf = (char *)g_static_private_get(&key_errbuf);
-
-	DLItem *item = (DLItem *)data;
-
-	curl_easy_setopt(curl, CURLOPT_URL, item->getUrl().c_str());
-	curl_easy_setopt(curl, CURLOPT_WRITEDATA, data);
-
-	LOG(LOG_DEBUG, "Starting to download %s", item->getUrl().c_str());
-
-	result = curl_easy_perform(curl);
-	if (result)
-	{
-		LOG(LOG_ERR, "Download failed for %s: %s",
-			item->getUrl().c_str(), errbuf);
-		/* XXX Handle the error */
-	}
-	else
-	{
-		LOG(LOG_DEBUG, "Finished download %s", item->getUrl().c_str());
-		/* XXX Check if the job is now ready */
-	}
+	g_static_mutex_unlock(&gc_lock);
 }
 
 /**********************************************************************
  * Class DownloadManager
  */
 
-DownloadManager::DownloadManager(int threads)
+DownloadManager::DownloadManager(int num_threads, int max_retries):max_retries(max_retries)
 {
-	pool = g_thread_pool_new(download_handler, NULL, threads, TRUE, NULL);
-	if (!pool)
+	do_exit = 0;
+
+	queue = g_queue_new();
+	queue_lock = g_mutex_new();
+	queue_sig = g_cond_new();
+
+	GError *error = NULL;
+	for (int i = 0; i < num_threads; i++)
 	{
-		LOG(LOG_ERR, "Failed to launch download threads");
-		exit(1);
+		GThread *thr = g_thread_create(DownloadManager::run_dl, this, TRUE, &error);
+		if (!thr)
+			throw QMException("Failed to create a new thread: %s", error->message);
+		threads.push_back(thr);
 	}
 
 	/* Initialize OpenSSL's thread interface */
@@ -163,18 +135,40 @@ DownloadManager::DownloadManager(int threads)
 
 	/* Initialize libcurl */
 	curl_global_init(CURL_GLOBAL_ALL);
-	shared_data = curl_share_init();
-	curl_share_setopt(shared_data, CURLSHOPT_LOCKFUNC, lock_curl);
-	curl_share_setopt(shared_data, CURLSHOPT_UNLOCKFUNC, unlock_curl);
-	curl_share_setopt(shared_data, CURLSHOPT_SHARE, CURL_LOCK_DATA_DNS);
+	shared_curl_data = curl_share_init();
+	shared_curl_lock = g_mutex_new();
+	curl_share_setopt(shared_curl_data, CURLSHOPT_LOCKFUNC, lock_curl);
+	curl_share_setopt(shared_curl_data, CURLSHOPT_UNLOCKFUNC, unlock_curl);
+	curl_share_setopt(shared_curl_data, CURLSHOPT_USERDATA, shared_curl_lock);
+	curl_share_setopt(shared_curl_data, CURLSHOPT_SHARE, CURL_LOCK_DATA_DNS);
 
 }
 
 DownloadManager::~DownloadManager()
 {
-	g_thread_pool_free(pool, TRUE, TRUE);
+	do_exit = 1;
 
-	curl_share_cleanup(shared_data);
+	/* Wake up all threads so they can exit */
+	g_mutex_lock(queue_lock);
+	g_cond_signal(queue_sig);
+	g_mutex_unlock(queue_lock);
+
+	LOG(LOG_DEBUG, "Waiting for downloads to finish");
+	while (!threads.empty())
+	{
+		GThread *thr = threads.front();
+		threads.erase(threads.begin());
+		g_thread_join(thr);
+	}
+
+	/* XXX Remove pending elements from the queue without calling either
+	 * ->finished() or ->failed() */
+	g_queue_free(queue);
+	g_mutex_free(queue_lock);
+	g_cond_free(queue_sig);
+
+	curl_share_cleanup(shared_curl_data);
+	g_mutex_free(shared_curl_lock);
 	curl_global_cleanup();
 
 	CRYPTO_set_id_callback(NULL);
@@ -186,6 +180,141 @@ DownloadManager::~DownloadManager()
 		g_free(ssl_mutexes);
 	}
 
+	run_gc();
+}
+
+static int sort_by_time(const void *a, const void *b, void *c G_GNUC_UNUSED)
+{
+	DLItem *aa = (DLItem *)a;
+	DLItem *bb = (DLItem *)b;
+
+	return *aa < *bb ? -1 : 1;
+}
+
+void DownloadManager::add(DLItem *item)
+{
+	g_mutex_lock(queue_lock);
+	g_queue_insert_sorted(queue, item, sort_by_time, NULL);
+	g_cond_signal(queue_sig);
+	g_mutex_unlock(queue_lock);
+}
+
+void DownloadManager::retry(DLItem *item)
+{
+	if (item->retries >= max_retries)
+	{
+		item->failed();
+		return;
+	}
+
+	g_get_current_time(&item->when);
+	g_time_val_add(&item->when, ++item->retries * 10000);
+
+	g_mutex_lock(queue_lock);
+	g_queue_insert_sorted(queue, item, sort_by_time, NULL);
+	g_cond_signal(queue_sig);
+	g_mutex_unlock(queue_lock);
+}
+
+static size_t write_callback(void *buf, size_t size, size_t items, void *data)
+{
+	DLItem *item = (DLItem *)data;
+
+	/* XXX Add support for pause/abort */
+
+	item->write(buf, size * items);
+	return size * items;
+}
+
+static int find_path(const void *a, const void *b)
+{
+	const DLItem *item = (const DLItem *)a;
+	const string *path = (const string *)b;
+
+	return !(item->getPath() == *path);
+}
+
+void DownloadManager::abort(const string &path)
+{
+	g_mutex_lock(queue_lock);
+	GList *link = g_queue_find_custom(queue, &path, find_path);
+	if (link)
+	{
+		DLItem *item = (DLItem *)link->data;
+		g_queue_delete_link(queue, link);
+		item->failed();
+	}
+	g_mutex_unlock(queue_lock);
+}
+
+void *DownloadManager::run_dl(void *data)
+{
+	DownloadManager *self = (DownloadManager *)data;
+	CURLcode result;
+	GTimeVal now;
+	char *errbuf;
+	CURL *curl;
+
+	curl = curl_easy_init();
+	curl_easy_setopt(curl, CURLOPT_SHARE, self->shared_curl_data);
+	errbuf = (char *)g_malloc(CURL_ERROR_SIZE);
+	curl_easy_setopt(curl, CURLOPT_ERRORBUFFER, errbuf);
+	curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_callback);
+	curl_easy_setopt(curl, CURLOPT_NOSIGNAL, (long)TRUE);
+	curl_easy_setopt(curl, CURLOPT_MAXREDIRS, 10);
+
+	/* XXX Add support for CURLOPT_SEEKFUNCTION to resume interrupted downloads */
+
+	/* XXX For debugging */
+	// curl_easy_setopt(curl, CURLOPT_VERBOSE, (long)TRUE);
+
+	LOG(LOG_DEBUG, "Download thread %p started", g_thread_self());
+
+	g_mutex_lock(self->queue_lock);
+
+	while (!self->do_exit)
+	{
+		g_get_current_time(&now);
+
+		/* Check if there is an item we can download */
+		DLItem *item = (DLItem *)g_queue_peek_head(self->queue);
+		if (!item || *item >= now)
+		{
+			if (item)
+				now = item->when;
+			g_cond_timed_wait(self->queue_sig, self->queue_lock, item ? &now : NULL);
+			continue;
+		}
+
+		/* Get the item off the queue and release the queue */
+		g_queue_pop_head(self->queue);
+		g_mutex_unlock(self->queue_lock);
+
+		curl_easy_setopt(curl, CURLOPT_URL, item->getUrl().c_str());
+		curl_easy_setopt(curl, CURLOPT_WRITEDATA, data);
+
+		LOG(LOG_DEBUG, "Starting to download %s", item->getUrl().c_str());
+
+		result = curl_easy_perform(curl);
+		if (result)
+		{
+			LOG(LOG_ERR, "Download failed for %s: %s", item->getUrl().c_str(), errbuf);
+			self->retry(item);
+		}
+		else
+		{
+			LOG(LOG_DEBUG, "Finished downloading %s", item->getUrl().c_str());
+			item->finished();
+		}
+		g_mutex_lock(self->queue_lock);
+	}
+
+	LOG(LOG_DEBUG, "Download thread %p exiting", g_thread_self());
+	g_mutex_unlock(self->queue_lock);
+
+	curl_easy_cleanup(curl);
+	g_free(errbuf);
+	return NULL;
 }
 
 /**********************************************************************
@@ -194,20 +323,75 @@ DownloadManager::~DownloadManager()
 
 DLItem::DLItem(const string &URL, const string &path):url(URL),path(path)
 {
-	fd = open(path.c_str(), O_WRONLY | O_CREAT | O_TRUNC, 0640);
-	if (fd == -1)
-		throw QMException("Failed to create file %s: %s",
-			path.c_str(), strerror(errno));
+	fd = -1;
+	g_get_current_time(&when);
+	retries = 0;
+
+	run_gc();
 }
 
 DLItem::~DLItem()
 {
 	if (fd != -1)
 		close(fd);
-	unlink(path.c_str());
 }
 
 size_t DLItem::write(void *buf, size_t size)
 {
+	if (fd == -1)
+	{
+		fd = open(path.c_str(), O_WRONLY | O_CREAT | O_TRUNC, 0640);
+		if (fd == -1)
+			throw QMException("Failed to create file %s: %s",
+				path.c_str(), strerror(errno));
+	}
 	return ::write(fd, buf, size);
+}
+
+void DLItem::finished()
+{
+	if (fd != -1)
+	{
+		close(fd);
+		fd = -1;
+	}
+	
+	add_to_gc(this);
+}
+
+void DLItem::failed()
+{
+	if (fd != -1)
+	{
+		close(fd);
+		fd = -1;
+	}
+	if (path.length())
+		unlink(path.c_str());
+
+	add_to_gc(this);
+}
+
+bool DLItem::operator<(const DLItem &b)
+{
+	return when.tv_sec < b.when.tv_sec ||
+		(when.tv_sec == b.when.tv_sec && when.tv_usec < b.when.tv_usec);
+}
+
+bool DLItem::operator<(const GTimeVal &b)
+{
+	return when.tv_sec < b.tv_sec ||
+		(when.tv_sec == b.tv_sec && when.tv_usec < b.tv_usec);
+}
+
+bool DLItem::operator>=(const GTimeVal &b)
+{
+	return when.tv_sec > b.tv_sec ||
+		(when.tv_sec == b.tv_sec && when.tv_usec >= b.tv_usec);
+}
+
+void DLItem::setRetry(const GTimeVal &when, int retries)
+{
+	this->when = when;
+	this->retries = retries;
 }
